updated core files, added portal page
