+++ app/page.module.scss
@@ -6,9 +6,26 @@
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
-  padding: 6rem;
-}
+  padding: 1rem;
+  
+  background-color: $cyd-light-gray;
 
-.test {
-  display: block;
+  .hero {
+    .hero-message {
+      @include desktop-m {
+        width: 50%;
+      }
+    }
+
+    .button-row {
+      display: grid;
+      gap: 1rem;
+      margin-top: 2rem;
+      @include desktop-s {
+        grid-template-columns: repeat(2, auto);
+        justify-items: center;
+        justify-content: start;
+      }
+    }
+  }
 }
