+++ app/portal/layout.js
@@ -1,0 +1,10 @@
+import Header from "components/layout/Header/Header"
+
+export default function PortalLayout({ children }) {
+  return (
+    <>
+      <Header />
+      {children}
+    </>
+  )
+}
