+++ app/portal/page.js
@@ -1,0 +1,12 @@
+import styles from "./page.module.scss"
+
+export default function Home() {
+  return (
+    // This is the landing page for users that are logged in
+    <main>
+      <section>
+        <p>You are logged in.</p>
+      </section>
+    </main>
+  )
+}
