+++ styles/_variables.scss
@@ -2,7 +2,8 @@
 // Note: in CSS use "black" and "white"
 $cyd-pink: #f33581;
 $cyd-dark-pink: #ca285e;
-$cyd-purple: #5A5CBF;
+$cyd-purple: #5a5cbf;
+$cyd-light-gray: #eff4f8;
 $cyd-gray: #dfdcd7;
 $cyd-black: #002a3e;
 // gradient colors
